+++ app/database/_bookshelfModels.js
@@ -1,0 +1,52 @@
+module.exports = function (bookshelf) {
+  const Authentication = bookshelf.Model.extend({
+    tableName: 'authentication',
+    hasTimeStamps: ['createdAt', 'updatedAt'],
+
+    user: function () {
+      return this.belongsTo(User, 'userId')
+    },
+  })
+
+  const Application = bookshelf.Model.extend({
+    tableName: 'applications',
+    hasTimeStamps: ['createdAt', 'updatedAt'],
+
+    user: function () {
+      return this.belongsTo(User, 'userId')
+    },
+    techPreferences: function () {
+      return this.hasMany(TechPreference, 'applicationId')
+    },
+  })
+
+  const TechPreference = bookshelf.Model.extend({
+    tableName: 'techpreferences',
+    hasTimeStamps: ['createdAt'],
+
+    application: function () {
+      return this.belongsTo(Application, 'applicationId')
+    },
+  })
+
+  const User = bookshelf.Model.extend({
+    tableName: 'users',
+    hasTimeStamps: ['createdAt', 'updatedAt'],
+
+    // relations
+    authentication () {
+      return this.hasMany(Authentication, 'userId')
+    },
+    application () {
+      return this.hasMany(Application, 'userId')
+    },
+  })
+
+  return {
+    bookshelf, // needed for stuff like `transaction`
+    Application,
+    Authentication,
+    TechPreference,
+    User,
+  }
+}
