+++ app/database/index.js
@@ -1,6 +1,6 @@
 const Knex = require('knex')
 
-const knexConfig = require('./knexConfig')
+const knexConfig = require('./_knexConfig')
 const knexInstance = Knex(knexConfig)
 const models = require('./models')(knexInstance)
