+++ app/logger.js
@@ -1,14 +1,14 @@
 const bunyan = require('bunyan')
 
-const logger = bunyan.createLogger({
-  name: 'hackcampus-apply',
-  serializers: bunyan.stdSerializers,
-})
+// const logger = bunyan.createLogger({
+//   name: 'hackcampus-apply',
+//   serializers: bunyan.stdSerializers,
+// })
 
-// const logger = {
-//   error: console.error,
-//   fatal: console.error,
-//   info: console.log,
-// }
+const logger = {
+  error: console.error,
+  fatal: console.error,
+  info: console.log,
+}
 
 module.exports = logger
